% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ContinentalUsEcoregion3}
\alias{ContinentalUsEcoregion3}
\title{Level 3 Ecoregions of the Continental United States}
\format{
An object of class \code{"sf"} with 1250 rows and 15 variables
}
\source{
\url{https://www.epa.gov/eco-research/level-iii-and-iv-ecoregions-continental-united-states}
}
\usage{
data(ContinentalUsEcoregion3)
}
\description{
The spatial boundaries, color symbology, and classification of ecoregions.
}
\references{
\href{ftp://newftp.epa.gov/EPADataCommons/ORD/Ecoregions/us/Eco_Level_III_descriptions.doc}{Description} \href{ftp://newftp.epa.gov/EPADataCommons/ORD/Ecoregions/us/Eco_Level_III_US.html}{Metadata}
}
\keyword{datasets}
