% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ContinentalUsEcoregion4}
\alias{ContinentalUsEcoregion4}
\title{Level 4 Ecoregions of the Continental United States}
\format{
An object of class \code{"sf"} with 5896 rows and 18 variables
}
\source{
\url{https://www.epa.gov/eco-research/level-iii-and-iv-ecoregions-continental-united-states}
}
\usage{
data(ContinentalUsEcoregion4)
}
\description{
The spatial boundaries, color symbology, and classification of ecoregions.
}
\references{
\href{ftp://newftp.epa.gov/EPADataCommons/ORD/Ecoregions/us/Eco_Level_IV_US.html}{Metadata}
}
\keyword{datasets}
