% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NorthAmericaEcoregion1}
\alias{NorthAmericaEcoregion1}
\title{Level 1 Ecoregions of North America}
\format{
An object of class \code{"sf"} with 2140 rows and 7 variables
}
\source{
\url{https://www.epa.gov/eco-research/ecoregions-north-america}
}
\usage{
data(NorthAmericaEcoregion1)
}
\description{
The spatial boundaries, color symbology, and classification of ecoregions.
}
\references{
\href{ftp://newftp.epa.gov/EPADataCommons/ORD/Ecoregions/cec_na/NA_CEC_Eco_Level1.htm}{Metadata}
}
\keyword{datasets}
