% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NorthAmericaEcoregion2}
\alias{NorthAmericaEcoregion2}
\title{Level 2 Ecoregions of North America}
\format{
An object of class \code{"sf"} with 2261 rows and 10 variables
}
\source{
\url{https://www.epa.gov/eco-research/ecoregions-north-america}
}
\usage{
data(NorthAmericaEcoregion2)
}
\description{
The spatial boundaries, color symbology, and classification of ecoregions.
}
\references{
\href{ftp://newftp.epa.gov/EPADataCommons/ORD/Ecoregions/cec_na/NA_CEC_Eco_Level2.htm}{Metadata}
}
\keyword{datasets}
