% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NorthAmericaEcoregion3}
\alias{NorthAmericaEcoregion3}
\title{Level 3 Ecoregions of North America}
\format{
An object of class \code{"sf"} with 2548 rows and 13 variables
}
\source{
\url{https://www.epa.gov/eco-research/ecoregions-north-america}
}
\usage{
data(NorthAmericaEcoregion3)
}
\description{
The spatial boundaries, color symbology, and classification of ecoregions.
}
\references{
\href{ftp://newftp.epa.gov/EPADataCommons/ORD/Ecoregions/pubs/NA_TerrestrialEcoregionsLevel3_Final-2june11_CEC.pdf}{Description} \href{ftp://newftp.epa.gov/EPADataCommons/ORD/Ecoregions/cec_na/NA_CEC_Eco_Level3.htm}{Metadata}
}
\keyword{datasets}
