% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worldecoregions}
\alias{worldecoregions}
\title{Terrestrial Ecoregions of the World}
\format{
An object of class \code{"sf"} with 14458 rows and 25 variables
}
\source{
\url{https://www.worldwildlife.org/publications/terrestrial-ecoregions-of-the-world}
}
\usage{
data(worldecoregions)
}
\description{
The spatial boundaries, color symbology, and classification of ecoregions.
}
\references{
Olson, D. M., Dinerstein, E., Wikramanayake, E. D., Burgess, N. D., Powell, G. V. N., Underwood, E. C., D'Amico, J. A., Itoua, I., Strand, H. E., Morrison, J. C., Loucks, C. J., Allnutt, T. F., Ricketts, T. H., Kura, Y., Lamoreux, J. F., Wettengel, W. W., Hedao, P., Kassem, K. R. 2001. Terrestrial ecoregions of the world: a new map of life on Earth. Bioscience 51(11):933-938.
}
\keyword{datasets}
